Remove old quotes concept
