+++ src/graphql/use-add-agreement-from-quote.tsx
@@ -4,11 +4,11 @@ import {
   AddAgreementFromQuoteMutationHookResult,
   AddAgreementFromQuoteMutationVariables,
   Contract,
+  GetQuotesDocument,
   Quote,
   useAddAgreementFromQuoteMutation,
 } from 'api/generated/graphql'
 import { format } from 'date-fns'
-import { QUOTES_QUERY } from './use-quotes'
 
 export const useAddAgreementFromQuote = (): AddAgreementFromQuoteMutationHookResult =>
   useAddAgreementFromQuoteMutation()
@@ -35,7 +35,7 @@ export const addAgreementFromQuoteOptions = (
     },
     refetchQueries: () => [
       {
-        query: QUOTES_QUERY,
+        query: GetQuotesDocument,
         variables: { memberId: contract.holderMemberId },
       },
     ],
