+++ src/graphql/use-create-quote.tsx
@@ -6,9 +6,9 @@ import {
   CreateQuoteFromAgreementMutationVariables,
   GenericAgreement,
   GetContractsDocument,
+  GetQuotesDocument,
   useCreateQuoteFromAgreementMutation,
 } from 'api/generated/graphql'
-import { QUOTES_QUERY } from './use-quotes'
 
 export const useCreateQuoteFromAgreement = (): CreateQuoteFromAgreementMutationHookResult =>
   useCreateQuoteFromAgreementMutation()
@@ -27,7 +27,7 @@ export const createQuoteFromAgreementOptions = (
     },
     refetchQueries: () => [
       {
-        query: QUOTES_QUERY,
+        query: GetQuotesDocument,
         variables: { memberId: contract.holderMemberId },
       },
       {
