+++ src/graphql/use-get-quotes.tsx
@@ -3,8 +3,12 @@ import {
   Quote,
   useGetQuotesQuery,
 } from 'api/generated/graphql'
+import { parseISO } from 'date-fns'
 
 type GetQuotesReturnTuple = [ReadonlyArray<Quote>, GetQuotesQueryHookResult]
+
+const latest = (a: Quote, b: Quote) =>
+  Number(parseISO(b.createdAt)) - Number(parseISO(a.createdAt))
 
 export const useQuotes = (memberId): GetQuotesReturnTuple => {
   const queryResult = useGetQuotesQuery({
@@ -12,8 +16,7 @@ export const useQuotes = (memberId): GetQuotesReturnTuple => {
       memberId,
     },
   })
-  const quotes = (queryResult.data?.member?.quotes ?? []) as ReadonlyArray<
-    Quote
-  >
-  return [quotes, queryResult]
+  const quotes = [...(queryResult.data?.member?.quotes ?? [])] as Quote[]
+
+  return [quotes.sort(latest), queryResult]
 }
