+++ src/graphql/use-sign-quote-for-new-contract.tsx
@@ -1,13 +1,13 @@
 import { MutationFunctionOptions } from '@apollo/react-common'
 import {
   GetContractsDocument,
+  GetQuotesDocument,
   SignQuoteForNewContractMutation,
   SignQuoteForNewContractMutationHookResult,
   SignQuoteForNewContractMutationVariables,
   useSignQuoteForNewContractMutation,
 } from 'api/generated/graphql'
 import { format } from 'date-fns'
-import { QUOTES_QUERY } from './use-quotes'
 
 export const useSignQuoteForNewContract = (): SignQuoteForNewContractMutationHookResult =>
   useSignQuoteForNewContractMutation()
@@ -27,7 +27,7 @@ export const getSignQuoteForNewContractOptions = (
     },
     refetchQueries: () => [
       {
-        query: QUOTES_QUERY,
+        query: GetQuotesDocument,
         variables: { memberId },
       },
       {
